+++ apps/chat-service/src/features/video-call/libs/attachSessionListeners.ts
@@ -1,6 +1,7 @@
 import { type Room, RoomEvent } from 'livekit-client';
 
 import type { VideoCallStateType } from '../model/types';
+import type { TrackInfo } from '../model/types';
 
 type StateType = Omit<VideoCallStateType, 'updateVideoCallState'>;
 
@@ -17,6 +18,7 @@ export const attachSessionListeners = (
                 {
                     trackPublication: publication,
                     participantIdentity: participant.identity,
+                    name: participant?.name || 'Anonymous',
                 },
             ],
         }));
@@ -43,6 +45,26 @@ export const attachSessionListeners = (
         updateVideoCallState({
             isConnected: true,
         });
+
+        const initialRemoteTracks: TrackInfo[] = [];
+
+        room.remoteParticipants.forEach((participant) => {
+            participant.trackPublications.forEach((publication) => {
+                if (publication.isSubscribed && publication.kind === 'video') {
+                    initialRemoteTracks.push({
+                        trackPublication: publication,
+                        participantIdentity: participant.identity,
+                        name: participant.name || 'Anonymous',
+                    });
+                }
+            });
+        });
+
+        if (initialRemoteTracks.length > 0) {
+            updateVideoCallState((prev) => ({
+                remoteTracks: [...prev.remoteTracks, ...initialRemoteTracks],
+            }));
+        }
     });
 
     room.on(RoomEvent.Disconnected, () => {
