+++ apps/chat-service/src/features/video-call/model/types.ts
@@ -7,6 +7,7 @@ import type {
 export interface TrackInfo {
     trackPublication: RemoteTrackPublication;
     participantIdentity: string;
+    name: string;
 }
 
 type StateType = {
