+++ apps/chat-service/src/features/video-call/ui/ChatMobileMessageHeader.tsx
@@ -11,27 +11,27 @@ export default function ChatMobileMessageHeader({
     chatRoomUuid,
     handlePrevStep,
 }: ChatMobileMessageHeaderProps) {
-    if (type !== 'message' || !chatRoomUuid) {
+    if (type === 'message' && chatRoomUuid) {
         return (
-            <div className='p-4 border-b border-gray-300 flex justify-center'>
+            <div className='p-4 border-b border-gray-300 flex justify-between'>
+                <button
+                    type='button'
+                    onClick={handlePrevStep}
+                    className='cursor-pointer text-gray-500 hover:text-gray-700 focus:text-gray-700'
+                >
+                    <Back />
+                </button>
+
                 <h3>채팅방</h3>
+
+                <div />
             </div>
         );
     }
 
     return (
-        <div className='p-4 border-b border-gray-300 flex justify-between'>
-            <button
-                type='button'
-                onClick={handlePrevStep}
-                className='cursor-pointer text-gray-500 hover:text-gray-700 focus:text-gray-700'
-            >
-                <Back />
-            </button>
-
+        <div className='p-4 border-b border-gray-300 flex justify-center'>
             <h3>채팅 목록</h3>
-
-            <div />
         </div>
     );
 }
