+++ apps/chat-service/src/features/video-call/ui/VideoRoom.tsx
@@ -5,8 +5,8 @@ import {
     SubscriberVideo,
     VideoControls,
 } from '@/features/video-call/ui';
-import SharedScreen from './SharedScreen';
 import { cn } from '@repo/ui/lib/utils';
+// import SharedScreen from './SharedScreen';
 
 export default function VideoRoom({ className }: { className?: string }) {
     const { isConnected } = useVideoCallContext();
@@ -16,13 +16,13 @@ export default function VideoRoom({ className }: { className?: string }) {
     return (
         <div
             className={cn(
-                'flex flex-col flex-wrap gap-4 h-full w-full p-6',
+                'flex flex-col flex-wrap gap-4 h-full w-full p-6 pr-0',
                 className,
             )}
         >
             <PublisherVideo />
             <SubscriberVideo />
-            <SharedScreen />
+            {/* <SharedScreen /> */}
 
             <VideoControls />
         </div>
